+++ src/main/kotlin/cloud/glitchdev/rfu/events/managers/TrophyCatchEvents.kt
@@ -12,8 +12,8 @@ import cloud.glitchdev.rfu.constants.fishing.TrophyFrog
 
 @AutoRegister
 object TrophyCatchEvents : RegisteredEvent {
-    private val TROPHY_FROG_REGEX = """ TROPHY FROG! You caught (?:an? )?(.+?) (BRONZE|SILVER|GOLD|DIAMOND)!""".toRegex(RegexOption.IGNORE_CASE)
-    private val TROPHY_FISH_REGEX = """ TROPHY FISH! You caught (?:an? )?(.+?) (BRONZE|SILVER|GOLD|DIAMOND)!""".toRegex(RegexOption.IGNORE_CASE)
+    val TROPHY_FROG_REGEX = """ TROPHY FROG! You caught (?:an? )?(.+?) (BRONZE|SILVER|GOLD|DIAMOND)!""".toRegex(RegexOption.IGNORE_CASE)
+    val TROPHY_FISH_REGEX = """ TROPHY FISH! You caught (?:an? )?(.+?) (BRONZE|SILVER|GOLD|DIAMOND)!""".toRegex(RegexOption.IGNORE_CASE)
 
     override fun register() {
         registerGameEvent(TROPHY_FROG_REGEX) { _, _, matches ->